requirements.txt
alembic==1.13.1
astroid==3.0.2
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
livereload==2.6.3
Mako==1.3.2
MarkupSafe==2.1.4
mccabe==0.7.0
mysql-connector-python==8.3.0
pep8==1.7.1
platformdirs==4.1.0
pylint==3.0.3
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
tomlkit==0.12.3
tornado==6.2
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
